+++ reconcilers/old.go
@@ -39,17 +39,24 @@ func (r *OldReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.R
 		return ctrl.Result{}, err
 	}
 
-	if r.migrator.ShouldMigrateSource(object) {
-		logger.Info("object should be migrated, skipping reconciliation")
-		return ctrl.Result{}, nil
-	}
-
 	syncer := r.syncerFactory.Syncer(object)
+
+	var sets appsv1.StatefulSetList
+	err = r.ctl.List(ctx, object.GetNamespace(), &sets, client.MatchingLabels(render.OldOwnershipLabels(object)))
+	if err != nil {
+		return ctrl.Result{}, err
+	}
 
 	if object.GetDeletionTimestamp() != nil {
 		if controllerutil.RemoveFinalizer(object, Finalizer) {
 			if _, err := syncer.DeleteAll(ctx); err != nil {
 				return ctrl.Result{}, err
+			}
+
+			for _, set := range sets.Items {
+				if err := r.ctl.Delete(ctx, &set); err != nil {
+					return ctrl.Result{}, err
+				}
 			}
 
 			if err := r.ctl.Update(ctx, object); err != nil {
@@ -69,13 +76,12 @@ func (r *OldReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.R
 		return ctrl.Result{}, nil
 	}
 
-	if _, err := syncer.Sync(ctx); err != nil {
-		return ctrl.Result{}, err
+	if r.migrator.ShouldMigrateSource(object) {
+		logger.Info("object should be migrated, skipping reconciliation")
+		return ctrl.Result{}, nil
 	}
 
-	var sets appsv1.StatefulSetList
-	err = r.ctl.List(ctx, object.GetNamespace(), &sets, client.MatchingLabels(render.OldOwnershipLabels(object)))
-	if err != nil {
+	if _, err := syncer.Sync(ctx); err != nil {
 		return ctrl.Result{}, err
 	}
 
@@ -85,7 +91,11 @@ func (r *OldReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.R
 		if err != nil {
 			return ctrl.Result{}, err
 		}
-		if outdated != nil {
+		stable, err := r.manager.PodsStable(ctx, &set)
+		if err != nil {
+			return ctrl.Result{}, err
+		}
+		if outdated != nil && stable {
 			logger.Info("found outdated pod, rolling", "pod", outdated.Name)
 			if err := r.ctl.Delete(ctx, outdated); err != nil {
 				return ctrl.Result{}, err
@@ -95,6 +105,9 @@ func (r *OldReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.R
 	}
 
 	desired := render.OldStatefulSet(r.operator, object)
+	if err := controllerutil.SetControllerReference(object, desired, r.ctl.Scheme()); err != nil {
+		return ctrl.Result{}, err
+	}
 	if err := r.ctl.Apply(ctx, desired, client.ForceOwnership); err != nil {
 		return ctrl.Result{}, err
 	}
